Umis 006

+++ app/Http/Controllers/UmisAndEmployeeManagement/ChildController.php
@@ -4,17 +4,13 @@ namespace App\Http\Controllers\UmisAndEmployeeManagement;
 
 use App\Http\Controllers\Controller;
 
-use Carbon\Carbon;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
-use Illuminate\Support\Str;
-use Illuminate\Support\Facades\Log;
 use App\Services\RequestLogger;
 use App\Http\Requests\ChildRequest;
 use App\Http\Resources\ChildResource;
 use App\Models\Child;
 use App\Models\EmployeeProfile;
-use App\Models\SystemLogs;
 
 class ChildController extends Controller
 {
@@ -34,14 +30,17 @@ class ChildController extends Controller
         try{
             $children = Child::where('personal_information_id', $id)->get();
 
-            if(!$child)
+            if(!$children)
             {
                 return response()->json(['message' => 'No record found.'], Response::HTTP_NOT_FOUND);
             }
 
             $this->requestLogger->registerSystemLogs($request, $children['id'], true, 'Success in fetching '.$this->PLURAL_MODULE_NAME.'.');
 
-            return response()->json(['data' => ChildResource::collection($children)], Response::HTTP_OK);
+            return response()->json([
+                'data' => ChildResource::collection($children),
+                'message' => 'Employee children record retrieved.'
+            ], Response::HTTP_OK);
         }catch(\Throwable $th){
             $this->requestLogger->errorLog($this->CONTROLLER_NAME,'findByPersonalInformationID', $th->getMessage());
             return response()->json(['message' => $th->getMessage()], Response::HTTP_INTERNAL_SERVER_ERROR);
@@ -63,7 +62,10 @@ class ChildController extends Controller
 
             $this->requestLogger->registerSystemLogs($request, $children['id'], true, 'Success in fetching '.$this->PLURAL_MODULE_NAME.'.');
 
-            return response()->json(['data' => ChildResource::collection($children)], Response::HTTP_OK);
+            return response()->json([,
+                'data' => ChildResource::collection($children),
+                'message' => 'Employee children record retrieved.'
+            ], Response::HTTP_OK);
         }catch(\Throwable $th){
             $this->requestLogger->errorLog($this->CONTROLLER_NAME,'findByEmployeeID', $th->getMessage());
             return response()->json(['message' => $th->getMessage()], Response::HTTP_INTERNAL_SERVER_ERROR);
@@ -87,7 +89,10 @@ class ChildController extends Controller
 
             $this->requestLogger->registerSystemLogs($request, $child['id'], true, 'Success in creating '.$this->SINGULAR_MODULE_NAME.'.');
 
-            return response()->json(['data' => 'Success'], Response::HTTP_OK);
+            return response()->json([
+                'data' => new ChildResource($child),
+                'message' => 'New employee child record added.'
+            ], Response::HTTP_OK);
         }catch(\Throwable $th){
             $this->requestLogger->errorLog($this->CONTROLLER_NAME,'store', $th->getMessage());
             return response()->json(['message' => $th->getMessage()], Response::HTTP_INTERNAL_SERVER_ERROR);
@@ -106,7 +111,10 @@ class ChildController extends Controller
 
             $this->requestLogger->registerSystemLogs($request, $id, true, 'Success in fetching '.$this->SINGULAR_MODULE_NAME.'.');
 
-            return response()->json(['data' => $child], Response::HTTP_OK);
+            return response()->json([
+                'data' => new ChildResource($child),
+                'message' => 'Employee child record retrieved'
+            ], Response::HTTP_OK);
         }catch(\Throwable $th){
             $this->requestLogger->errorLog($this->CONTROLLER_NAME,'show', $th->getMessage());
             return response()->json(['message' => $th->getMessage()], Response::HTTP_INTERNAL_SERVER_ERROR);
@@ -137,7 +145,10 @@ class ChildController extends Controller
 
             $this->requestLogger->registerSystemLogs($request, $id, true, 'Success in updating '.$this->SINGULAR_MODULE_NAME.'.');
 
-            return response()->json(['data' => new ChildResource($child), 'message' => 'Employee child data is updated.'], Response::HTTP_OK);
+            return response()->json([
+                'data' => new ChildResource($child), 
+                'message' => 'Employee child data is updated.'
+            ], Response::HTTP_OK);
         }catch(\Throwable $th){
             $this->requestLogger->errorLog($this->CONTROLLER_NAME,'update', $th->getMessage());
             return response()->json(['message' => $th->getMessage()], Response::HTTP_INTERNAL_SERVER_ERROR);
@@ -170,12 +181,12 @@ class ChildController extends Controller
         try{
             $children = Child::where('personal_information_id', $id)->get();
 
-            if(!$children)
+            if(count($children) === 0)
             {
                 return response()->json(['message' => 'No record found.'], Response::HTTP_NOT_FOUND);
             }
 
-            foreach($child as $key => $child){
+            foreach($children as $key => $child){
                 $child->delete();
             }
             
@@ -201,11 +212,16 @@ class ChildController extends Controller
             $personal_information = $employee_profile->personalInformation;
             $children = $personal_information->children;
 
-            foreach($child as $key => $child){
+            if(count($children) === 0)
+            {
+                return response()->json(['message' => 'No record found.'], Response::HTTP_NOT_FOUND);
+            }
+
+            foreach($children as $key => $child){
                 $child->delete();
             }
             
-            $this->requestLogger->registerSystemLogs($request, $id, true, 'Success in deleting '.$this->SINGULAR_MODULE_NAME.'.');
+            $this->requestLogger->registerSystemLogs($request, null, true, 'Success in deleting '.$this->SINGULAR_MODULE_NAME.'.');
             
             return response()->json(['message' => 'Employee children record deleted'], Response::HTTP_OK);
         }catch(\Throwable $th){
